archive/cleanup_20251029/requirements_freeze_before_4.5.12.txt
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.10
aiosignal==1.4.0
alembic==1.13.3
annotated-types==0.7.0
astroid==4.0.1
asttokens==3.0.0
asyncpg==0.29.0
attrs==25.3.0
Authlib==1.6.5
bandit==1.7.10
black==24.10.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
ccxt==4.4.8
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
circuitbreaker==2.0.0
click==8.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
coverage==7.3.3
cryptography==42.0.8
dateparser==1.2.2
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
dparse==0.6.4
execnet==2.1.1
executing==2.2.1
factory-boy==3.3.0
Faker==20.1.0
filelock==3.12.4
flake8==7.1.1
Flask==3.1.2
Flask-BasicAuth==0.2.0
flask-cors==6.0.1
frozenlist==1.7.0
gevent==25.9.1
geventhttpclient==2.3.4
greenlet==3.2.4
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.1.0
ipython==8.18.1
isort==7.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
locust==2.17.0
loguru==0.7.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mirakuru==2.6.1
msgpack==1.1.1
multidict==6.6.4
mypy==1.11.2
mypy_extensions==1.1.0
numpy==2.1.1
opentelemetry-api==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-aiohttp-client==0.48b0
opentelemetry-instrumentation-asyncpg==0.48b0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
packaging==25.0
pandas==2.2.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.4.0
pluggy==1.6.0
port-for==0.7.4
prometheus_client==0.21.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==6.0.0
psycopg==3.2.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybit==5.8.0
pycares==4.11.0
pycodestyle==2.12.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.19.2
pylint==4.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-benchmark==4.0.0
pytest-cov==5.0.0
pytest-flake8==1.1.1
pytest-mock==3.14.0
pytest-mypy==0.10.3
pytest-postgresql==5.0.0
pytest-timeout==2.3.1
pytest-xdist==3.5.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
redis==6.4.0
regex==2025.9.18
requests==2.32.5
rich==14.1.0
roundrobin==0.0.4
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safety==3.2.8
safety-schemas==0.0.16
scipy==1.14.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.35
stack-data==0.6.3
stevedore==5.5.0
structlog==24.4.0
tenacity==9.0.0
tomlkit==0.13.3
traitlets==5.14.3
typer==0.19.2
types-cachetools==5.5.0.20240820
types-PyYAML==6.0.12.20240917
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvloop==0.20.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
